Conflicts:
	wsitransformer/pom.xml

+++ iTopologyManager/topologyViewerConfigAPI/src/main/java/net/itransformers/topologyviewer/config/models/datamatcher/impl/LessOrEqual.java
@@ -46,8 +46,8 @@ public class LessOrEqual implements DataMatcher{
     @Override
     public boolean compareData(String data1, String data2) {
         try {
-            Integer i1 = Integer.parseInt(data1);
-            Integer i2 = Integer.parseInt(data2);
+            Long i1 = Long.parseLong(data1);
+            Long i2 = Long.parseLong(data2);
             logger.debug("LorEq: "+i1+"<="+i2);
             return i1 <= i2;
         }catch (RuntimeException rte) {
